+++ AudioBooth/AudioBooth/Services/UserPreferences/UserPreferences+Cloud.swift
@@ -64,6 +64,8 @@ extension UserPreferences {
     "podcastEpisodeFilter",
     "podcastEpisodeSort",
     "podcastEpisodeSortAscending",
+    "podcastAutoQueueSettings",
+    "autoDownloadQueuedEpisodes",
     "openPlayerOnLaunch",
     "playbackSpeed",
     "speedPresets",